first crack at html2markdown and back.
