+++ vue/package.json
@@ -33,6 +33,7 @@
     "promise.prototype.finally": "^3.1.2",
     "raven-js": "latest",
     "register-service-worker": "^1.6.2",
+    "showdown": "^1.9.1",
     "svg.js": "^2.7.1",
     "tippy.js": "^4.3.5",
     "tiptap": "1.26.6",
