+++ vue/src/shared/services/format_converter.coffee
@@ -1,0 +1,14 @@
+import marked from 'marked'
+import {customRenderer, options} from '@/shared/helpers/marked.coffee'
+marked.setOptions Object.assign({renderer: customRenderer()}, options)
+
+export convertToHtml = (model, field) ->
+  model["#{field}Format"] = "html"
+  model[field] = marked(model[field] || '')
+
+export convertToMd = (model, field) ->
+  import("showdown").then (showdown) =>
+    console.log showdown
+    converter = new showdown.Converter()
+    model["#{field}Format"] = "md"
+    model[field] = converter.makeMarkdown(model[field] || '')
